Move decoration into TraceContext
